Basic multiprofile awareness in Security settings
The various lock- and biometric-related settings in Security will now
no longer assume a single work profile. This requires passing the
work profile ID to the involved controllers.

Launch the Security activity of Settings from a specific profile in
order to change such settings for that profile. The behavior of the
Settings app in the parent user remains unchanged, only allowing
access to the settings for a single work profile.

Tests *may* currently fail.

Change-Id: I87f036787a325d51be04cce14d7a5ec125329ccb