Basic multiprofile awareness in Security settings
These changes should have *no functional impact* unless a user starts
the Security activity of Settings from a child profile, which is not
easily possible without using adb or an app that can browse and start
arbitrary app activities. The behavior of the Settings app in the
parent user remains unchanged, only allowing access to the settings
for a single profile.

The various lock- and biometric-related preferences in Security will
now no longer assume a single child profile. To accomplish this, we
pass the profile ID to the involved preference controllers.

Prior to this change, some settings would apply to the parent user,
or fail to be properly applied at all, when the Security activity
was started from a child profile. Now, lock settings affect the ID
that started the activity, allowing existing Settings UI elements
in the Security activity to be utilized within a profile.

Impacts at least:
* Lock challenge (PIN/password/etc.)
* Biometric (Fingerprint, Face Unlock)
* Power Button Instantly Locks
* Screen Timeout

TODO: Verify whether these settings are effective when changed from
the Work Profile section vs the "main user" section (which now is
meant to be treated as the "whoever launched me" section).

Tests *may* currently fail.

Change-Id: I87f036787a325d51be04cce14d7a5ec125329ccb