Basic multiprofile awareness in Security settings
The various lock- and biometric-related settings in Security will now
no longer assume a single work profile. This requires passing the
work profile ID to the involved controllers.

In addition, prior to this change, some settings would apply to the
parent user, or fail to be properly applied at all, when the Security
activity was launched from a work profile. Now, lock settings change
for the user/profile that launched the activity, allowing existing
Settings UI elements for these settings to be utilized.

Impacts at least:
* Lock challenge (PIN/password/etc.)
* Biometric (Fingerprint, Face Unlock)
* Power Button Instantly Locks
* Screen Timeout

TODO: Verify whether these settings are effective when changed from
the Work Profile section vs the "main user" section (which now is
meant to be treated as the "whoever launched me" section).

Launch the Security activity of Settings from a specific profile in
order to change such settings for that profile. The behavior of the
Settings app in the parent user remains unchanged, only allowing
access to the settings for a single work profile.

Tests *may* currently fail.

Change-Id: I87f036787a325d51be04cce14d7a5ec125329ccb