Always use combined biometrics settings
Devices that only support one type of biometric authentication will
have their settings moved into a combined section now, too.

The combined biometrics settings activity provides more granular
control over the behavior of face and fingerprint unlock, including
whether or not they can be used to unlock the device or merely be
used to "Verify it's you in apps."

This change exposes AOSP's combined biometric section in Security,
or in the unified Security & Privacy section when enabled, which
contains settings for either or both biometric unlock options,
depending on hardware.

It also adjusts strings and text to be more appropriate for single-
method devices and adds a missing "Ways to use" title string.

Co-authored-by: Oliver Scott <olivercscott@gmail.com>
Issue: calyxos#1165
Change-Id: Ib188cabb762bd30036682f2e4248597b62e45cb0