Disable removing the first org-owned profile
Disable the removal button for the first managed profile when the
device is organization-owned - in other words, when the device was
set up with the Safer or Safest garlic levels. Removal is already
prevented at a framework level, but without this change, attempted
removal via Settings silently fails, leading to user confusion.
Profile owner apps like Bellis always have the ability to remove
a profile, so if applicable, a garlic-level profile can be removed
from the Bellis app.

Note: In base AOSP, work profiles on organization-owned devices cannot
be removed at all, but we allow it in CalyxOS so that users utilizing
multiple profiles are not stuck with profiles they no longer want.

Change-Id: I86e7613bb02e7cd12807ff1c16d53ee5a929feba