MAYBE DO NOT MERGE: rebase match: FingerprintEnrollEnrolling
The android-13.0.0_r16 merge added a theme parameter for the alert
dialog, which (maybe accidentally) didn't get used by us. We also (maybe
accidentally) started setting the title instead of the message. This
change matches 4.12.2 in those behaviors. This should not be merged if
those were accidental; otherwise, squash with:
FingerprintEnrollEnrolling: Set proper dialog message when user touches fp icon

(This also includes a newline at the end of the file that isn't in
AOSP.)

Change-Id: Iedd62038f43ac696ddac3f101cb615df2f5fed23