MAYBE DO NOT MERGE: rebase match: FingerprintEnrollEnrolling
AOSP added a theme parameter for the alert dialog, but we held it back
in 4.12.2, probably on accident during a merge. This change matches 4.12.2
but should not be merged if it was just an accident.

Otherwise, if it was on purpose or if this doesn't work, squash with:
FingerprintEnrollEnrolling: Set proper dialog message when user touches fp icon

Change-Id: Iedd62038f43ac696ddac3f101cb615df2f5fed23