Notify that global cleartext restriction is removed if enabled
If global cleartext restriction was enabled, notify that it has been
removed. We already warned previously, so this alters the warning
to become a new and final notification confirming that it is now
removed.

Additionally, this notification only shows for the device's main user
so that it won't show for every new guest or user created in the future,
which would happen unless we removed this deleted setting from the
database, and also so it won't be unnecessarily duplicated for work
or private profiles.

Issue: calyxos#2987
Change-Id: I471dc92998af463be428881c6d203eec4e246602